refactor: move codes of switching processes to `switch` module
